backend/requirements.txt
# Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.6.1
chromadb>=0.5.0
